Move DynamicAnalysisResults struct to public pkg/ directory
